Rename AudioProcessor to ModularProcessor
This relfects the reality that AudioProcessors are not always processing
audio. They can process other things like MIDI inputs, and the only only
thing that really unites them is the fact that they can recieve data
from parameters.

Also, the parameter registration system was moved from being handled in
CommonPlugBase to in the ModularProcessor. App code in the old system
was a little confusing between RegisterProcessor() and
RegisterProcessorParameter() that did very different things.